CNDIT-1153 addressing PR comments, moving TestUtils to common-util module
