+++ data-reporting-service/common-util/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
+    implementation 'commons-io:commons-io:2.15.0'
     implementation 'org.springframework.kafka:spring-kafka:3.1.2'
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.17.0'
     implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
