+++ data-reporting-service/common-util/src/main/java/gov/cdc/etldatapipeline/commonutil/TestUtils.java
@@ -1,4 +1,4 @@
-package gov.cdc.etldatapipeline.observation;
+package gov.cdc.etldatapipeline.commonutil;
 
 import org.apache.commons.io.FileUtils;
 import org.springframework.core.io.ClassPathResource;
