+++ data-reporting-service/investigation-service/src/main/java/gov/cdc/etldatapipeline/investigation/controller/InvestigationController.java
@@ -3,17 +3,19 @@ package gov.cdc.etldatapipeline.investigation.controller;
 import gov.cdc.etldatapipeline.investigation.service.KafkaProducerService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 
 @RestController
 @RequiredArgsConstructor
 @Slf4j
 public class InvestigationController {
     private final KafkaProducerService producerService;
-    private final String topicName = "nbs_Investigation";
+
+    @Value("${spring.kafka.stream.input.investigation.topic-name}")
+    private String topicName;
 
     @GetMapping("/reporting/investigation-svc/status")
     @ResponseBody
