--- data-reporting-service/investigation-service/src/test/java/gov/cdc/etldatapipeline/investigation/TestUtils.java
@@ -1,20 +1,0 @@
-package gov.cdc.etldatapipeline.investigation;
-
-import org.apache.commons.io.FileUtils;
-import org.springframework.core.io.ClassPathResource;
-
-import java.io.IOException;
-import java.nio.charset.Charset;
-
-public class TestUtils {
-
-    public static String readFileData(String fileName) {
-        try {
-            return FileUtils.readFileToString(
-                    new ClassPathResource(fileName).getFile(),
-                    Charset.defaultCharset());
-        } catch (IOException e) {
-            throw new RuntimeException("File Read failed : " + fileName);
-        }
-    }
-}
