+++ data-reporting-service/organization-service/src/test/java/gov/cdc/etldatapipeline/organization/DebeziumChangeDataParserTests.java
@@ -4,7 +4,7 @@ import gov.cdc.etldatapipeline.organization.model.odse.Organization;
 import gov.cdc.etldatapipeline.organization.utils.UtilHelper;
 import org.junit.jupiter.api.Test;
 
-import static gov.cdc.etldatapipeline.organization.TestUtils.readFileData;
+import static gov.cdc.etldatapipeline.commonutil.TestUtils.readFileData;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class DebeziumChangeDataParserTests {
