+++ data-reporting-service/person-service/src/test/java/gov/cdc/etldatapipeline/person/PersonDetailsDeserializationTests.java
@@ -5,7 +5,7 @@ import gov.cdc.etldatapipeline.person.model.dto.persondetail.*;
 import gov.cdc.etldatapipeline.person.utils.UtilHelper;
 import org.junit.jupiter.api.Test;
 
-import static gov.cdc.etldatapipeline.person.TestUtils.readFileData;
+import static gov.cdc.etldatapipeline.commonutil.TestUtils.readFileData;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PersonDetailsDeserializationTests {
