+++ data-reporting-service/person-service/src/test/java/gov/cdc/etldatapipeline/person/config/PersonSvcYmlPropertiesTest.java
@@ -23,9 +23,9 @@ public class PersonSvcYmlPropertiesTest {
     void whenBindingYMLConfigFile_thenAllFieldsAreSet() {
         Assertions.assertEquals("nbs_default", kafkaConfig.getDefaultDataTopicName());
         Assertions.assertEquals("nbs_Person", kafkaConfig.getPersonTopicName());
-        Assertions.assertEquals("elastic_nrt_patient", kafkaConfig.getPatientElasticSearchTopic());
+        Assertions.assertEquals("elastic_search_patient", kafkaConfig.getPatientElasticSearchTopic());
         Assertions.assertEquals("nrt_patient", kafkaConfig.getPatientReportingTopic());
-        Assertions.assertEquals("elastic_nrt_provider", kafkaConfig.getProviderElasticSearchTopic());
+        Assertions.assertEquals("elastic_search_provider", kafkaConfig.getProviderElasticSearchTopic());
         Assertions.assertEquals("nrt_provider", kafkaConfig.getProviderReportingTopic());
     }
